requirements.txt
flask==3.0.2
python-decouple==3.8
openai==1.82.1 